requirements.txt
prometheus-client
koyeb-api-client
requests
